Removed TorchServe Also added the ability to bypass LitServe
