+++ backend/PythonPackages/setup/dev-requirements.txt
@@ -5,3 +5,5 @@
 opencv-python~=4.11.0.86
 matplotlib~=3.10.0
 termcolor~=2.5.0
+
+-r ext-requirements.txt
