--- backend/PythonPackages/setup/ts-requirements.txt
@@ -1,6 +1,0 @@
-# TorchServe dependencies
-
-torchserve~=0.12.0
-torch-model-archiver~=0.12.0
-pyyaml~=6.0.2
-captum~=0.7.0
