--- model_server/.gitignore
@@ -1,2 +1,0 @@
-model_store/
-logs/
