--- model_server/config.properties
@@ -1,5 +1,0 @@
-inference_address=http://127.0.0.1:6060
-management_address=http://127.0.0.1:6061
-metrics_address=http://127.0.0.1:6062
-model_store=model_store/
-disable_token_authorization=true
